+++ tests/packagecoverage.R
@@ -10,17 +10,9 @@
 #}
 
 #test_for_me("R/evaluate.R", "tests/testthat/test-evaluate.R")
-  
 # Look at whole package -------------------------------------------------------
-
 # Gets percent coverage for entire package, by script
-  
 #covr::package_coverage()
 #
 # Report with gui (probably the best here)
 #covr::report()
-
-
-
-
-