+++ NEWS.md
@@ -1,5 +1,7 @@
 # TBSignatureProfiler (development version)
 
+# TBSignatureProfiler 1.92.0
+
 ## Bug Fixes
 * Fixed the tableAUC bootstrapped confidence interval to be the 2.5 and 97.5 percentiles instead of the 5 and 95 percentiles
 * Fixed the upper CI value for the pROC/DeLong AUC CI method in the bootstrapAUC function
@@ -10,6 +12,7 @@
 ## Minor Changes
 * Updated the github and website introductions.
 * Added Natarajan_7, Kaul_3 signatures
+* Added Francisco_OD_2, Kwan_186 signatures
 * Shortened some example run times
 
 # TBSignatureProfiler 1.8.0